remove engine and replace functionality with peer
